Refactor asset imports and restructure components for improved organization
