--- src/assetts.js
@@ -1,30 +1,0 @@
-import shape136 from "../src/assets/shape136.png";
-import checkSign from "./assets/checkSign.png";
-import ww from "./assets/ww.png";
-import rightarrow from "./assets/rightarrow.png";
-import qq from "./assets/shape_138.png";
-import nn from "./assets/shape_137.png";
-import zz from "./assets/shape_172.png";
-import SPCFZ from "./assets/SPCFZ.png";
-import expand from "./assets/expandArrow.png";
-import collapse from "./assets/collapseArrow.png";
-import linkedin from "./assets/linkedin.png";
-import facebook from "./assets/facebook.png";
-import iconsClose from "./assets/iconsClose.png";
-import iconsmMenu from "./assets/iconsmMenu.png";
-export default {
-  ww,
-  shape136,
-  qq,
-  zz,
-  nn,
-  SPCFZ,
-  collapse,
-  expand,
-  checkSign,
-  rightarrow,
-  linkedin,
-  facebook,
-  iconsClose,
-  iconsmMenu,
-};
