+++ src/components/Accordation/Accordion.jsx
@@ -1,5 +1,5 @@
 import React, { Children, useState } from "react";
-import assetts from "../assetts";
+import assetts from "../../JsData/assetts";
 import { motion } from "framer-motion";
 export default function Accordion({ title, children }) {
   const [isOpen, setIsOpen] = useState();
