+++ src/components/whyUs/WhyUs.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import assetts from "../../assetts.js";
+import assetts from "../../JsData/assetts.js";
 import { motion, useInView } from "framer-motion";
 import { useRef } from "react";
